refactoring, move http middleware to separate location
