+++ http/middleware/errors.go
@@ -1,7 +1,7 @@
 // Copyright © 2020 by PACE Telematics GmbH. All rights reserved.
 // Created at 2020/08/27 by Marius Neugebauer
 
-package http
+package middleware
 
 import "errors"
